move MDX component map to root
