+++ apps/www/src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Route } from 'react-router-dom'
 import { MDXProvider } from '@mdx-js/tag'
 import '@cwds/core/dist/styles.css'
 import { Utils, JumpToTop } from '@cwds/components'
-import componentMap from './components/MDXComponentProviderMappings'
+import componentMap from './MDXComponentProviderMappings'
 
 import ScrollToTop from './ScrollToTop'
 import Status from './modules/status'
@@ -29,16 +29,15 @@ class App extends Component {
           <Router>
             <ScrollToTop>
               <Fragment>
-                <Route exact path="/" component={Home} />
-                <Route path="/core-style" component={CoreStyles} />
-                <Route path="/labs" component={LabsPage} />
-                <Route path="/components" component={Components} />
-                <Route path="/status" component={Status} />
-                <JumpToTop />
-              </Fragment>
-            </ScrollToTop>
-          </Router>
-        </CaresProvider>
+                <Route exact path="/" component={Home} />{' '}
+                <Route path="/core-style" component={CoreStyles} />{' '}
+                <Route path="/labs" component={LabsPage} />{' '}
+                <Route path="/components" component={Components} />{' '}
+                <Route path="/status" component={Status} /> <JumpToTop />
+              </Fragment>{' '}
+            </ScrollToTop>{' '}
+          </Router>{' '}
+        </CaresProvider>{' '}
       </MDXProvider>
     )
   }
